+++ src/main/scala/com/atomist/param/AllowedValue.scala
@@ -2,5 +2,5 @@ package com.atomist.param
 
 import com.fasterxml.jackson.annotation.{JsonCreator, JsonProperty}
 
-case class AllowedValue @JsonCreator()(@JsonProperty("name") name: String,
+case class AllowedValue @JsonCreator()(@JsonProperty("value") value: String,
                                        @JsonProperty("display_name") displayName: String)