+++ src/test/scala/com/atomist/rug/runtime/js/JavaScriptInvokingProjectOperationTest.scala
@@ -1,20 +1,20 @@
 package com.atomist.rug.runtime.js
 
 import com.atomist.project.SimpleProjectOperationArguments
-import com.atomist.rug.{InvalidRugParameterPatternException, TestUtils}
+import com.atomist.rug.{InvalidRugParameterDefaultValue, InvalidRugParameterPatternException, TestUtils}
 import com.atomist.source.{FileArtifact, SimpleFileBasedArtifactSource, StringFileArtifact}
 import org.scalatest.{FlatSpec, Matchers}
 
 object JavaScriptInvokingProjectOperationTest {
 
-  val SimpleEditorInvokingOtherEditorAndAddingToOurOwnParameters =
+  val SimpleEditorInvokingOtherEditorAndAddingToOurOwnParameters: String =
     s"""
        |import {Project} from '@atomist/rug/model/Core'
        |import {ProjectEditor} from '@atomist/rug/operations/ProjectEditor'
        |import {File} from '@atomist/rug/model/Core'
        |import {Result,Status,Parameter} from '@atomist/rug/operations/RugOperation'
        |
-           |class SimpleEditor implements ProjectEditor {
+       |class SimpleEditor implements ProjectEditor {
        |    name: string = "Simple"
        |    description: string = "A nice little editor"
        |    parameters: Parameter[] = [{name: "content", description: "Content", pattern: "@url", maxLength: 100}]
@@ -27,7 +27,7 @@ object JavaScriptInvokingProjectOperationTest {
        |var editor = new SimpleEditor()
     """.stripMargin
 
-  val SimpleReviewerInvokingOtherEditorAndAddingToOurOwnParameters =
+  val SimpleReviewerInvokingOtherEditorAndAddingToOurOwnParameters: String =
     s"""
        |import {Project} from '@atomist/rug/model/Core'
        |import {ProjectReviewer} from '@atomist/rug/operations/ProjectReviewer'
@@ -47,7 +47,7 @@ object JavaScriptInvokingProjectOperationTest {
        |var reviewer = new SimpleReviewer()
     """.stripMargin
 
-  val SimpleEditorWithBrokenParameterPattern =
+  val SimpleEditorWithBrokenParameterPattern: String =
       s"""
          |import {Project} from '@atomist/rug/model/Core'
          |import {ProjectEditor} from '@atomist/rug/operations/ProjectEditor'
@@ -68,7 +68,7 @@ object JavaScriptInvokingProjectOperationTest {
          |var editor = new SimpleEditor()
     """.stripMargin
 
-  val SimpleReviewerWithBrokenParameterPattern =
+  val SimpleReviewerWithBrokenParameterPattern: String =
     s"""
        |import {Project} from '@atomist/rug/model/Core'
        |import {ProjectReviewer} from '@atomist/rug/operations/ProjectReviewer'
@@ -86,6 +86,90 @@ object JavaScriptInvokingProjectOperationTest {
        |    }
        |  }
        |var reviewer = new SimpleReviewer()
+    """.stripMargin
+
+  val SimpleEditorWithInvalidDefaultParameterValuePattern: String =
+    s"""
+       |import {Project} from '@atomist/rug/model/Core'
+       |import {ProjectEditor} from '@atomist/rug/operations/ProjectEditor'
+       |import {File} from '@atomist/rug/model/Core'
+       |import {Result,Status,Parameter} from '@atomist/rug/operations/RugOperation'
+       |
+       |class SimpleEditor implements ProjectEditor {
+       |    name: string = "Simple"
+       |    description: string = "A nice little editor"
+       |    parameters: Parameter[] = [
+       |      {
+       |        name: "content",
+       |        description: "Content",
+       |        pattern: "@url",
+       |        default: "not-a-url",
+       |        maxLength: 100
+       |      }
+       |    ]
+       |    edit(project: Project, {content} : {content: string}) {
+       |      return new Result(Status.Success,
+       |        `Edited Project now containing $${project.fileCount()} files: \n`
+       |      );
+       |    }
+       |  }
+       |var editor = new SimpleEditor()
+    """.stripMargin
+
+  val SimpleEditorWithValidDefaultParameterValueFromAlternation: String =
+    s"""
+       |import {Project} from '@atomist/rug/model/Core'
+       |import {ProjectEditor} from '@atomist/rug/operations/ProjectEditor'
+       |import {File} from '@atomist/rug/model/Core'
+       |import {Result,Status,Parameter} from '@atomist/rug/operations/RugOperation'
+       |
+       |class SimpleEditor implements ProjectEditor {
+       |    name: string = "Simple"
+       |    description: string = "A nice little editor"
+       |    parameters: Parameter[] = [
+       |      {
+       |        name: "content",
+       |        description: "Content",
+       |        pattern: "^(?:http://a.b.c|http://g.co|ftp://f.co)$$",
+       |        default: "http://g.co",
+       |        maxLength: 100
+       |      }
+       |    ]
+       |    edit(project: Project, {content} : {content: string}) {
+       |      return new Result(Status.Success,
+       |        `Edited Project now containing $${project.fileCount()} files: \n`
+       |      );
+       |    }
+       |  }
+       |var editor = new SimpleEditor()
+    """.stripMargin
+
+  val SimpleEditorWithInvalidDefaultParameterValueAlternation: String =
+    s"""
+       |import {Project} from '@atomist/rug/model/Core'
+       |import {ProjectEditor} from '@atomist/rug/operations/ProjectEditor'
+       |import {File} from '@atomist/rug/model/Core'
+       |import {Result,Status,Parameter} from '@atomist/rug/operations/RugOperation'
+       |
+       |class SimpleEditor implements ProjectEditor {
+       |    name: string = "Simple"
+       |    description: string = "A nice little editor"
+       |    parameters: Parameter[] = [
+       |      {
+       |        name: "content",
+       |        description: "Content",
+       |        pattern: "^(?:http://a.b.c|http://g.co|ftp://f.co)$$",
+       |        default: "http://g.com",
+       |        maxLength: 100
+       |      }
+       |    ]
+       |    edit(project: Project, {content} : {content: string}) {
+       |      return new Result(Status.Success,
+       |        `Edited Project now containing $${project.fileCount()} files: \n`
+       |      );
+       |    }
+       |  }
+       |var editor = new SimpleEditor()
     """.stripMargin
 }
 
@@ -113,6 +197,22 @@ class JavaScriptInvokingProjectOperationTest extends FlatSpec with Matchers {
     }
   }
 
+  it should "run simple editor and throw an exception for default parameter value not matching pattern" in {
+    assertThrows[InvalidRugParameterDefaultValue] {
+      invokeAndVerifyEditorWithDefaults(StringFileArtifact(s".atomist/reviewers/SimpleEditor.ts", SimpleEditorWithInvalidDefaultParameterValuePattern))
+    }
+  }
+
+  it should "run simple editor compiled from TypeScript and validate the default from allowed values correctly" in {
+    invokeAndVerifyEditorWithDefaults(StringFileArtifact(s".atomist/reviewers/SimpleEditor.ts", SimpleEditorWithValidDefaultParameterValueFromAlternation))
+  }
+
+  it should "run simple editor and throw an exception for default parameter value not in list" in {
+    assertThrows[InvalidRugParameterDefaultValue] {
+      invokeAndVerifyEditorWithDefaults(StringFileArtifact(s".atomist/reviewers/SimpleEditor.ts", SimpleEditorWithInvalidDefaultParameterValueAlternation))
+    }
+  }
+
   private def invokeAndVerifySimpleEditor(tsf: FileArtifact): JavaScriptInvokingProjectEditor = {
     val as = TestUtils.compileWithModel(SimpleFileBasedArtifactSource(tsf))
     val jsed = JavaScriptOperationFinder.fromJavaScriptArchive(as).head.asInstanceOf[JavaScriptInvokingProjectEditor]
@@ -133,5 +233,16 @@ class JavaScriptInvokingProjectOperationTest extends FlatSpec with Matchers {
       case _ =>
     }
     jsr
+  }
+
+  private def invokeAndVerifyEditorWithDefaults(tsf: FileArtifact): JavaScriptInvokingProjectEditor = {
+    val as = TestUtils.compileWithModel(SimpleFileBasedArtifactSource(tsf))
+    val jsed = JavaScriptOperationFinder.fromJavaScriptArchive(as).head.asInstanceOf[JavaScriptInvokingProjectEditor]
+    jsed.name should be("Simple")
+    val target = SimpleFileBasedArtifactSource(StringFileArtifact("pom.xml", "nasty stuff"))
+    jsed.modify(target, SimpleProjectOperationArguments("", Map[String,String]())) match {
+      case _ =>
+    }
+    jsed
   }
 }